supervisor_profiles_api.conf bug fixed

+++ deploy/setup.sh
@@ -36,7 +36,7 @@ supervisorctl restart profiles_api
 
 # Configure nginx
 cp $PROJECT_BASE_PATH/deploy/nginx_profiles_api.conf /etc/nginx/sites-available/profiles_api.conf
-rm /etc/nginx/sites-enabled/default
+[ -d "/etc/nginx/sites-enabled/default" ] || rm /etc/nginx/sites-enabled/default
 ln -s /etc/nginx/sites-available/profiles_api.conf /etc/nginx/sites-enabled/profiles_api.conf
 systemctl restart nginx.service
